some refactoring
"разгрузил" MainActivity.kt
