+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
             android:name="com.example.effective_lab.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.AndroidLab">
+            android:theme="@style/Theme.Effective_labs">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
