+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">AndroidLab</string>
+    <string name="app_name">effective_labs</string>
 </resources>
